refactor: extract view models and split SecurityService
Extract 37 view model data classes from WebPageFactory.kt into
ViewModels.kt. WebPageFactory now contains only the factory class
and utility functions.

Split SecurityService into focused services:
- PasswordResetService: password reset request/confirm
- ApiKeyService: API key creation, auth, listing, deletion
- OAuthService: OAuth user lookup/creation

SecurityService delegates to these services, keeping its public
API intact so no callers need updating.

Verified: 505 tests pass, all quality checks clean.
